Add support for **immediately** rendering a given show
This adds a new obnoxious link per show that causes a given show to be displayed immediately.

It piggybacks on the existing OSC functionality for doing that, and is exactly as janky, if not slightly more. It's great

Note: If you have any ideas for a better name than "[...]_immediately_show_show", we're more than happy to change it. Also sorry.